Bugfix on InputVoltage example

+++ Examples/InputVoltage/main.cpp
@@ -13,10 +13,10 @@ void setup() {
 void loop(){
 
     //Read the board input voltage,
-    float voltage = Loka::getInputVoltage();
+    double voltage = Loka::getInputVoltage();
 
     //Print the value
-    console_debug("Input Voltage: %f", voltage);
+    console_debug("Input Voltage: %lf", voltage);
 
     //Sleep for 1 second
     delay(1000);
